[opentitantool] Extend HyperDebug driver timeout
Many interactions with the HyperDebug hardware is in the form of sending
console commands via USB, and waiting for an acknowledgement/response.

Traditionally the timeout for receiving such responses was 100ms.  But
we are seeing some indication, that this may not be enough in all cases.

In the happy case, increasing this timeout has no effect, as the driver
will stop reading, as soon as it has seen the next prompt character, and
the timeout will not be reached.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I468515476dacac7f0e4044b818a7d6159b30723f