[opentitantool] Avoid repeated HyperDebug communication
The recently introduced `SpiWrapper` has to set speed and chip select on
the underlying SPI bus before each transaction, because it cannot rule
out the possibility of some other `SpiWrapper` having operated on the
same underlying SPI bus in the meantime.

Before this CL, this would translate into many repeated commands asking
HyperDebug to set the speed to the same value.

This CL catches this in the HyperDebug transport driver, by keeping
track of which speed it has most recently told HyperDebug to use, for
each SPI bus.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ic0e6ccc1e0dc2470c7adc377b1dc9e1d4a03f74d